requirements.txt
requests
pandas
numpy
scikit-learn
sentence-transformers
streamlit
pyyaml
sqlalchemy
python-dotenv
chromadb
schedule
streamlit 
tqdm
